refactor(web): move realtime from shared to infra
