+++ turbo/apps/web/src/lib/infra/run/executors/runner-executor.ts
@@ -13,7 +13,7 @@ import {
 import { encryptSecretsMap } from "../../../shared/crypto/secrets-encryption";
 import { isOfficialRunnerGroup } from "../runner-group";
 import { forbidden } from "../../../shared/errors";
-import { publishJobNotification } from "../../../shared/realtime/client";
+import { publishJobNotification } from "../../realtime/client";
 import { logger } from "../../../shared/logger";
 import { recordSandboxOperation } from "../../../shared/metrics";
 import type { PreparedContext, ExecutorResult } from "./types";
